% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model.R
\name{as_rpdb_assembly}
\alias{as_rpdb_assembly}
\title{Convert Data to an rPDBapi Assembly Object}
\usage{
as_rpdb_assembly(x, metadata = list())
}
\arguments{
\item{x}{Input data.}

\item{metadata}{Optional metadata list.}
}
\value{
Object of class \code{rPDBapi_assembly}.
}
\description{
Convert Data to an rPDBapi Assembly Object
}
