% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model.R
\name{as_rpdb_chemical_component}
\alias{as_rpdb_chemical_component}
\title{Convert Data to an rPDBapi Chemical Component Object}
\usage{
as_rpdb_chemical_component(x, metadata = list())
}
\arguments{
\item{x}{Input data.}

\item{metadata}{Optional metadata list.}
}
\value{
Object of class \code{rPDBapi_chemical_component}.
}
\description{
Convert Data to an rPDBapi Chemical Component Object
}
