% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model.R
\name{as_rpdb_entry}
\alias{as_rpdb_entry}
\title{Convert Data to an rPDBapi Entry Object}
\usage{
as_rpdb_entry(x, metadata = list())
}
\arguments{
\item{x}{Input data.}

\item{metadata}{Optional metadata list.}
}
\value{
Object of class \code{rPDBapi_entry}.
}
\description{
Convert Data to an rPDBapi Entry Object
}
