% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model.R
\name{as_rpdb_polymer_entity}
\alias{as_rpdb_polymer_entity}
\title{Convert Data to an rPDBapi Polymer Entity Object}
\usage{
as_rpdb_polymer_entity(x, metadata = list())
}
\arguments{
\item{x}{Input data.}

\item{metadata}{Optional metadata list.}
}
\value{
Object of class \code{rPDBapi_polymer_entity}.
}
\description{
Convert Data to an rPDBapi Polymer Entity Object
}
