% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_model.R
\name{as_rpdb_structure}
\alias{as_rpdb_structure}
\title{Convert Data to an rPDBapi Structure Object}
\usage{
as_rpdb_structure(x, metadata = list())
}
\arguments{
\item{x}{Input data.}

\item{metadata}{Optional metadata list.}
}
\value{
Object of class \code{rPDBapi_structure}.
}
\description{
Convert Data to an rPDBapi Structure Object
}
