% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceOperator.R
\name{autoresolve_sequence_type}
\alias{autoresolve_sequence_type}
\title{Automatically Determine the Sequence Type}
\usage{
autoresolve_sequence_type(sequence)
}
\arguments{
\item{sequence}{A string representing the sequence to be analyzed.}
}
\value{
A string indicating the resolved sequence type.
}
\description{
This function determines the type of a given sequence (DNA, RNA, or PROTEIN) based on its characters.
}
