% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_helpers.R
\name{build_assembly_id}
\alias{build_assembly_id}
\title{Build an Assembly Identifier}
\usage{
build_assembly_id(entry_id, assembly_id)
}
\arguments{
\item{entry_id}{Entry identifier.}

\item{assembly_id}{Assembly index or identifier.}
}
\value{
Character scalar assembly ID formatted as \code{ENTRY-ASSEMBLY}.
}
\description{
Build an Assembly Identifier
}
