% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_helpers.R
\name{build_entity_id}
\alias{build_entity_id}
\title{Build an Entity Identifier}
\usage{
build_entity_id(entry_id, entity_id)
}
\arguments{
\item{entry_id}{Entry identifier.}

\item{entity_id}{Entity index or identifier.}
}
\value{
Character scalar entity ID formatted as \code{ENTRY_ENTITY}.
}
\description{
Build an Entity Identifier
}
