% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_helpers.R
\name{build_entry_id}
\alias{build_entry_id}
\title{Build an Entry Identifier}
\usage{
build_entry_id(entry_id)
}
\arguments{
\item{entry_id}{Entry identifier.}
}
\value{
Character scalar entry ID.
}
\description{
Build an Entry Identifier
}
