% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceOperator.R
\name{cannot_autoresolve_sequence_type_error}
\alias{cannot_autoresolve_sequence_type_error}
\title{Raise Error for Unresolvable Sequence Types}
\usage{
cannot_autoresolve_sequence_type_error(message)
}
\arguments{
\item{message}{A descriptive error message to display.}
}
\description{
Throws an error when a sequence type cannot be automatically determined.
This function is used within the autoresolve_sequence_type function.
}
