% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_client.R
\name{CannotInferSearchServiceException}
\alias{CannotInferSearchServiceException}
\title{Exception for Inability to Infer Search Service}
\usage{
CannotInferSearchServiceException(message)
}
\arguments{
\item{message}{A string containing the error message.}
}
\description{
Throws an exception when the search service cannot be inferred from the provided search operator.
}
