% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalOperator.R
\name{ChemicalOperator}
\alias{ChemicalOperator}
\title{Chemical Operator for SMILES/InChI Searches}
\usage{
ChemicalOperator(descriptor, matching_criterion = "graph-strict")
}
\arguments{
\item{descriptor}{A valid SMILES or InChI string.}

\item{matching_criterion}{The criterion for matching, one of the values from DescriptorMatchingCriterion.
Defaults to "graph-strict".}
}
\value{
A list representing a chemical operator.
}
\description{
Creates an object representing a chemical search operator using SMILES or InChI descriptors.
}
