% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{ComparisonOperator}
\alias{ComparisonOperator}
\title{Create a Comparison Search Operator}
\usage{
ComparisonOperator(attribute, value, comparison_type)
}
\arguments{
\item{attribute}{The attribute to be compared.}

\item{value}{The value to compare against.}

\item{comparison_type}{A string specifying the type of comparison (e.g., 'equal', 'greater_than').}
}
\value{
An object of class 'ComparisonOperator'.
}
\description{
Constructs a ComparisonOperator object for search operations that perform comparison checks on attribute values.
}
