% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{DefaultOperator}
\alias{DefaultOperator}
\title{Create a Default Search Operator}
\usage{
DefaultOperator(value)
}
\arguments{
\item{value}{The value to be used in the search operation.}
}
\value{
An object of class 'DefaultOperator' representing the default search operator.
}
\description{
Constructs a DefaultOperator object for use in general search operations within the RCSB PDB.
This operator is a basic search operator with a single value.
}
