% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{ExistsOperator}
\alias{ExistsOperator}
\title{Create an Existence Search Operator}
\usage{
ExistsOperator(attribute)
}
\arguments{
\item{attribute}{The attribute whose existence is to be checked.}
}
\value{
An object of class 'ExistsOperator'.
}
\description{
Constructs an ExistsOperator object for search operations to check the existence of an attribute.
}
