% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{extract_taxonomy_table}
\alias{extract_taxonomy_table}
\title{Extract Taxonomy Table}
\usage{
extract_taxonomy_table(x)
}
\arguments{
\item{x}{Polymer-entity-like table or object.}
}
\value{
Tibble with taxonomy columns when available.
}
\description{
Extract Taxonomy Table
}
