% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdb_api_url.R
\name{get_pdb_api_url}
\alias{get_pdb_api_url}
\title{Generate a PDB API URL}
\usage{
get_pdb_api_url(endpoint, id, base_url = BASE_URL)
}
\arguments{
\item{endpoint}{A character string representing the specific API endpoint to be accessed (e.g., "/pdb/entry/").}

\item{id}{A character string representing the identifier for the resource (e.g., a PDB ID or other relevant ID).}

\item{base_url}{A string representing the base URL to generate PDB API url. By default, this is set to the global constant \code{BASE_URL}, but users can specify a different URL if needed.}
}
\value{
A character string containing the full URL for accessing the PDB API resource.
}
\description{
This function constructs a full PDB API URL by concatenating the base URL, an API endpoint, and an identifier.
}
