% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdb_file.R
\name{get_pdb_file}
\alias{get_pdb_file}
\title{Download PDB Files from RCSB Database}
\usage{
get_pdb_file(
  pdb_id,
  filetype = "cif",
  rm.insert = FALSE,
  rm.alt = TRUE,
  compression = FALSE,
  save = FALSE,
  path = NULL
)
}
\arguments{
\item{pdb_id}{A 4-character string specifying a PDB entry of interest.}

\item{filetype}{The file type: 'pdb', 'cif', 'xml', or 'structfact'. Default is 'pdb'.pdb is the older file format and cif is the newer replacement. xlm is also can be obtained and parsed. structfact retrieves structure factors (only available for certain PDB entries).}

\item{rm.insert}{Logical, if TRUE PDB insert records are ignored.}

\item{rm.alt}{Logical, if TRUE PDB alternate records are ignored.}

\item{compression}{Logical indicating whether to request the data as a compressed version. Default is FALSE.}

\item{save}{Logical, if TRUE saves PDB file to the desired path.}

\item{path}{The path where the file should be saved. If NULL, the file is saved in a temporary directory.}
}
\value{
The uncompressed string representing the full PDB file or NULL if the request fails.
}
\description{
This function downloads PDB files from the RCSB database, supporting different file types and optional compression.
It warns to consider compression for CIF files for faster download.
}
