% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_client.R
\name{infer_search_service}
\alias{infer_search_service}
\title{Infer the Appropriate Search Service}
\usage{
infer_search_service(search_operator)
}
\arguments{
\item{search_operator}{A query operator object.}
}
\value{
The inferred search service.
}
\description{
Determines the appropriate search service for a given search operator in RCSB PDB queries.
}
