% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{InOperator}
\alias{InOperator}
\title{Create an Inclusion Search Operator}
\usage{
InOperator(attribute, value)
}
\arguments{
\item{attribute}{The attribute to be evaluated.}

\item{value}{The set of values to include in the search.}
}
\value{
An object of class 'InOperator'.
}
\description{
Constructs an InOperator object for search operations where the attribute value must be within a specified set.
}
