% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta_text_to_list.R
\name{parse_fasta_text_to_list}
\alias{parse_fasta_text_to_list}
\title{Helper Function: Parse FASTA Text to List Grouped by Header}
\usage{
parse_fasta_text_to_list(fasta_text)
}
\arguments{
\item{fasta_text}{A string containing FASTA-formatted text.}
}
\value{
A list where each element is a FASTA sequence keyed by the header.
}
\description{
This function parses a FASTA-formatted text into a list of sequences,
where each sequence is keyed by the header (which includes organism name and chain information).
}
