% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{RangeOperator}
\alias{RangeOperator}
\title{Create a Range Search Operator}
\usage{
RangeOperator(
  attribute,
  from_value,
  to_value,
  include_lower = TRUE,
  include_upper = TRUE,
  negation = FALSE
)
}
\arguments{
\item{attribute}{The attribute to be evaluated within a range.}

\item{from_value}{The starting value of the range.}

\item{to_value}{The ending value of the range.}

\item{include_lower}{Boolean to include the lower bound in the range.}

\item{include_upper}{Boolean to include the upper bound in the range.}

\item{negation}{Boolean to negate the range condition.}
}
\value{
An object of class 'RangeOperator'.
}
\description{
Constructs a RangeOperator object for search operations that specify a range for attribute values.
}
