% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_client.R
\name{RequestOptions}
\alias{RequestOptions}
\title{Define Options for Search Requests}
\usage{
RequestOptions(
  result_start_index = NULL,
  num_results = NULL,
  sort_by = "score",
  desc = TRUE
)
}
\arguments{
\item{result_start_index}{An integer specifying the starting index for result pagination, default is NULL.}

\item{num_results}{An integer specifying the number of results to return, default is NULL.}

\item{sort_by}{A string indicating the attribute to sort by, default is 'score'.}

\item{desc}{A boolean indicating whether sorting should be in descending order, default is TRUE.}
}
\value{
A list of request options.
}
\description{
Sets various options for RCSB PDB search requests, such as pagination and sorting.
}
