% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_id.R
\name{validate_id}
\alias{validate_id}
\title{Validate the Format of an PDB ID Based on Data Type}
\usage{
validate_id(id, data_type)
}
\arguments{
\item{id}{A string representing the ID to be validated.}

\item{data_type}{A string specifying the data type against which the ID is validated
(e.g., 'entity', 'instance', 'ASSEMBLY').}
}
\description{
This function checks if the given PDB ID is valid for the specified data type.
It issues warnings if the ID format does not match the expected format for the data type.
}
\details{
The function checks for specific patterns in the ID based on the data type:
  - For data types containing 'entity' but not 'instance', IDs should contain an underscore ('_').
  - For data types containing 'instance', IDs should contain a dot ('.').
  - For 'ASSEMBLY' data type, IDs should contain a dash ('-').
}
