% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{ContainsPhraseOperator}
\alias{ContainsPhraseOperator}
\title{Create a Contains Phrase Search Operator}
\usage{
ContainsPhraseOperator(attribute, value)
}
\arguments{
\item{attribute}{The attribute to be evaluated.}

\item{value}{The phrase to search for in the attribute.}
}
\value{
An object of class 'ContainsPhraseOperator'.
}
\description{
Constructs a ContainsPhraseOperator object for search operations that look for attributes containing a specific phrase.
}
