% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_chemical.R
\name{describe_chemical}
\alias{describe_chemical}
\title{Describe Chemical Compound from RCSB PDB}
\usage{
describe_chemical(chem_id)
}
\arguments{
\item{chem_id}{A string representing the 3-character chemical ID.}
}
\value{
A dictionary containing the chemical description.
}
\description{
This function retrieves the description of a chemical compound from the RCSB PDB based on its ID.
}
\examples{
chem_desc <- describe_chemical('NAG')
print(chem_desc$rcsb_chem_comp_descriptor$smiles)
}
