% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOperators.R
\name{ExactMatchOperator}
\alias{ExactMatchOperator}
\title{Create an Exact Match Search Operator}
\usage{
ExactMatchOperator(attribute, value)
}
\arguments{
\item{attribute}{The attribute to match.}

\item{value}{The exact value to search for.}
}
\value{
An object of class 'ExactMatchOperator'.
}
\description{
Constructs an ExactMatchOperator object for precise search operations in the RCSB PDB.
It matches an exact attribute value.
}
