% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_data_as_dataframe.R
\name{return_data_as_dataframe}
\alias{return_data_as_dataframe}
\title{Transform Response Data into a Dataframe}
\usage{
return_data_as_dataframe(response, data_type, ids)
}
\arguments{
\item{response}{A list containing the response data from a PDB query.}

\item{data_type}{A string indicating the type of data contained in the response.
This parameter is not directly used in the function but might be relevant for context.}

\item{ids}{A vector of identifiers corresponding to the response data.}
}
\value{
A dataframe constructed from the response data.
  Returns NULL if the response is null or empty.
}
\description{
This function converts response data from the RCSB Protein Data Bank (PDB) into a structured dataframe.
It handles null or empty responses and flattens the data for dataframe conversion.
}
